test(sqrt): fix int -> uint64_t sign-conversion warning in perf info
verify_heavy lambda 内部累加器是 uint64_t, 但 for-loop 变量是 int. 编译器
触发 -Wsign-conversion. 改成 uint64_t 循环变量, 无行为变化, warning 消除.

+++ tests/test_couveignes_parallel.cpp
@@ -408,7 +408,7 @@ void test_perf_info() {
     auto verify_heavy = [](uint64_t pattern) {
         // Simulate moderate verify cost with a tight integer-only loop.
         uint64_t acc = pattern;
-        for (int i = 0; i < 1000; ++i) {
+        for (uint64_t i = 0; i < 1000; ++i) {
             acc = acc * 2654435761ULL + i;
         }
         // Match at a single late position so both N=1 and N=4 do most of
